% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmexample1.R
\docType{data}
\name{cpmexample1}
\alias{cpmexample1}
\title{Dataset for the CPM method}
\format{
A data frame composed of predetermined columns:
\describe{
  \item{from}{starting activity node}
  \item{to}{final activity node}
  \item{label}{activity label}
  \item{time}{duration of the activity}
}
}
\usage{
cpmexample1
}
\description{
Fictitious data that is used in the examples.
In this dataset, the activities occur on the edges
}
\keyword{datasets}
