% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmexample2.R
\docType{data}
\name{cpmexample2}
\alias{cpmexample2}
\title{Dataset for the CPM method}
\format{
A data frame composed of predetermined columns:
\describe{
  \item{label}{activity label}
  \item{pred}{preceding activities}
  \item{time}{duration of the activity}
}
}
\usage{
cpmexample2
}
\description{
Example from  Miszczyńska D., Miszczyński M. "Wybrane metody badań operacyjnych" (2000, ISBN:83-907712-0-9).
10 activities, 8 nodes.
In this dataset, the activities occur on the edges and a list of direct predecessors has been added.
}
\keyword{datasets}
