% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lessexample1.R
\docType{data}
\name{lessexample1}
\alias{lessexample1}
\title{Dataset for the LESS method}
\format{
A data frame composed of predetermined columns:
\describe{
  \item{from}{starting activity node}
  \item{to}{final activity node}
  \item{label}{activity label}
  \item{time}{normal duration of the activity}
  \item{bound_time}{the shortest duration of the activity}
  \item{norm_cost}{normal cost of the activity}
  \item{bound_cost}{boundary cost of the activity}
}
}
\usage{
lessexample1
}
\description{
Fictitious data that is used in the examples.
In this dataset, the activities occur on the edges
}
\keyword{datasets}
