% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lessexample2.R
\docType{data}
\name{lessexample2}
\alias{lessexample2}
\title{Dataset for the LESS method}
\format{
A data frame composed of predetermined columns:
\describe{
  \item{label}{activity label}
  \item{pred}{preceding activities}
  \item{time}{normal duration of the activity}
  \item{bound_time}{the shortest duration of the activity}
  \item{norm_cost}{normal cost of the activity}
  \item{bound_cost}{boundary cost of the activity}
}
}
\usage{
lessexample2
}
\description{
Example from  Miszczyńska D., Miszczyński M. "Wybrane metody badań operacyjnych" (2000, ISBN:83-907712-0-9).
In this dataset, the activities occur on the edges and a list of direct predecessors has been added.
}
\keyword{datasets}
