% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath_methodsAOE.R
\name{PERT_newprob}
\alias{PERT_newprob}
\title{Probability for the given directive term}
\usage{
PERT_newprob(new_DT, yourlist)
}
\arguments{
\item{new_DT}{The given project completion date. The parameter must be greater than zero.}

\item{yourlist}{List of objects that make up the solution to the project management problem.}
}
\value{
This function calculates the probability of completing the project within the time specified by the user. A normal distribution was assumed.
}
\description{
Probability for the given directive term
}
\examples{
y <- solve_pathAOA(pertexample1, deterministic = FALSE)
PERT_newprob(new_DT = 30, y)
}
