% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath_methodsAOE.R
\name{PERT_newtime}
\alias{PERT_newtime}
\title{A new directive term for any probability}
\usage{
PERT_newtime(new_prob = 0.5, yourlist)
}
\arguments{
\item{new_prob}{Probability of the project completion. Default set to 0.5.}

\item{yourlist}{List of objects that make up the solution to the project management problem.}
}
\value{
This function computes a new directive term for a probability given by the user. A normal distribution was assumed.
}
\description{
A new directive term for any probability
}
\examples{
y <- solve_pathAOA(pertexample1, deterministic = FALSE)
PERT_newtime(new_prob = 0.3, y)
}
