% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pertexample1.R
\docType{data}
\name{pertexample1}
\alias{pertexample1}
\title{Dataset for the CPM method}
\format{
A data frame composed of predetermined columns:
\describe{
  \item{from}{starting activity node}
  \item{to}{final activity node}
  \item{label}{activity label}
  \item{opt_time}{optimistic duration of activity}
  \item{likely_time}{the most likely duration of the activity}
  \item{pes_time}{pesimistic duration of activity}
}
}
\usage{
pertexample1
}
\description{
Fictitious data that is used in the examples.
In this dataset, the activities occur on the edges
}
\keyword{datasets}
