% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pertexample2.R
\docType{data}
\name{pertexample2}
\alias{pertexample2}
\title{Dataset for the PERT method}
\format{
A data frame composed of predetermined columns:
\describe{
  \item{label}{activity label}
  \item{pred}{preceding activities}
  \item{opt_time}{optimistic duration of activity}
  \item{likely_time}{the most likely duration of the activity}
  \item{pes_time}{pesimistic duration of activity}
}
}
\usage{
pertexample2
}
\description{
Example from  Miszczyńska D., Miszczyński M. "Wybrane metody badań operacyjnych" (2000, ISBN:83-907712-0-9).
10 activities, 8 nodes.
In this dataset, the activities occur on the edges and a list of direct predecessors has been added.
}
\keyword{datasets}
