% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath_methodsAOE.R
\name{plot_crit_pathAOA}
\alias{plot_crit_pathAOA}
\title{Graph with marked critical path}
\usage{
plot_crit_pathAOA(yourlist)
}
\arguments{
\item{yourlist}{Lista obiektow tworzacych rozwiazanie problemu zarzadzania projektem}
}
\value{
The function draws the graph along with the critical path by means of the DiagrammeR package functions.
}
\description{
Graph with marked critical path
}
\examples{
x <- solve_pathAOA(cpmexample1, TRUE)
plot_crit_pathAOA(x)
}
