% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath_methodsAOE.R
\name{plot_gantt}
\alias{plot_gantt}
\title{A Gantt chart}
\usage{
plot_gantt(yourlist, bar_size = 10)
}
\arguments{
\item{yourlist}{List of objects that make up the solution to the project management problem.}

\item{bar_size}{Thickness of the bar drawn for activity (set to 10 by default).}
}
\value{
Draws a Gantt chart broken down into critical ("CR") and non-critical ("NC") activities.
  Marks the slack of time.
}
\description{
A Gantt chart
}
\examples{
x <- solve_pathAOA(cpmexample1, deterministic = TRUE)
plot_gantt(x)
}
