% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath_methodsAOE.R
\name{plot_graphAOA}
\alias{plot_graphAOA}
\title{Graph without critical path}
\usage{
plot_graphAOA(input_data)
}
\arguments{
\item{input_data}{Data frame describing the problem}
}
\value{
The function draws a relationship graph between activities without solving the problem
  and thus without marking critical activities.
}
\description{
Graph without critical path
}
\examples{
plot_graphAOA(cpmexample1)
}
