% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath_methodsAOE.R
\name{plot_norm}
\alias{plot_norm}
\title{The cumulative distribution function of the normal distribution}
\usage{
plot_norm(yourlist)
}
\arguments{
\item{yourlist}{List of objects making up the solution to the project management problem}
}
\value{
Draws a graph of the normal distribution with the expected directive term from
  the PERT method and the standard deviation for this term. The chart also includes lines indicating
  the schedules of the risk-taker and the belayer.
}
\description{
The cumulative distribution function of the normal distribution
}
\examples{
y <- solve_pathAOA(pertexample1, deterministic = FALSE)
plot_norm(y)
}
