% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lessAOA.R
\name{plot_TC}
\alias{plot_TC}
\title{Total cost change plot}
\usage{
plot_TC(your_list)
}
\arguments{
\item{your_list}{List containing solved problem}
}
\value{
Based on the results of the LESS method, a graph of the total cost value of all iterations is created
}
\description{
Total cost change plot
}
\examples{
z <- solve_lessAOA(lessexample1, 50, 15)
plot_TC(z)
}
