% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticalpath_methodsAOE.R
\name{plot_asap}
\alias{plot_asap}
\title{An ASAP chart}
\usage{
plot_asap(yourlist, show_dummy = FALSE, bar_size = 10)
}
\arguments{
\item{yourlist}{List of objects that make up the solution to the project management problem.}

\item{show_dummy}{Decides whether dummy activities should be included in the chart. If so, set it to TRUE (set to FALSE by default).}

\item{bar_size}{Thickness of the bar drawn for activity (set to 10 by default).}
}
\value{
Draws an ASAP (activities start and finish As Soon As Possible) chart broken down into critical ("CR") and non-critical ("NC") activities.
  Marks total floats.
}
\description{
An ASAP chart
}
\examples{
x <- solve_pathAOA(cpmexample1, deterministic = TRUE)
plot_asap(x)
}
